fix: ロジックテストで発覚したバグ修正
